Arduino: Use one parameter instead of multiple
Up until now this server used a different parameter for
different
colors/relays. This meant that for setting a different color
to the LED
strip you needed to make up to 3 requests and 4 requests for
the relay.

This changes the system to use a single parameter, a 16-bit
integer. One
byte for each parameter (RGB and relay).